Manual properties functionality
Adds functionality to the previously added UI elements.
Manual properties can now be set by the user as personal annotations to the data. They are stored in the hdf5 file and can be exported to an Excel file.
Adds on to / Replaces the process of creating an Excel file and making annotations to the data by hand.

Future improvements:
-Code is messy and variable naming is inconsistent.
-Exceptionally well suited for object oriented approach but currently uses dicts and dispersed functions --> Less control over correctness of data.